+++ tests/Constraints/NumberAndIntegerTypesTest.php
@@ -12,10 +12,8 @@ class NumberAndIntegerTypesTest extends BaseTestCase
     public function getInvalidTests(): \Generator
     {
         yield [
-            '{
-              "integer": 1.4
-            }',
-            '{
+            'input' => '{ "integer": 1.4 }',
+            'schema' => '{
               "type":"object",
               "properties":{
                 "integer":{"type":"integer"}
@@ -23,8 +21,8 @@ class NumberAndIntegerTypesTest extends BaseTestCase
             }'
         ];
         yield [
-            '{"integer": 1.001}',
-            '{
+            'input' => '{"integer": 1.001}',
+            'schema' => '{
                 "type": "object",
                 "properties": {
                     "integer": {"type": "integer"}
@@ -32,8 +30,8 @@ class NumberAndIntegerTypesTest extends BaseTestCase
             }'
         ];
         yield [
-            '{"integer": true}',
-            '{
+            'input' => '{"integer": true}',
+            'schema' => '{
                 "type": "object",
                 "properties": {
                     "integer": {"type": "integer"}
@@ -41,8 +39,8 @@ class NumberAndIntegerTypesTest extends BaseTestCase
             }'
         ];
         yield [
-            '{"number": "x"}',
-            '{
+            'input' => '{"number": "x"}',
+            'schema' => '{
                 "type": "object",
                 "properties": {
                     "number": {"type": "number"}
@@ -54,10 +52,8 @@ class NumberAndIntegerTypesTest extends BaseTestCase
     public function getValidTests(): \Generator
     {
         yield [
-            '{
-              "integer": 1
-            }',
-            '{
+            'input' => '{ "integer": 1 }',
+            'schema' => '{
               "type":"object",
               "properties":{
                 "integer":{"type":"integer"}
@@ -65,10 +61,8 @@ class NumberAndIntegerTypesTest extends BaseTestCase
             }'
         ];
         yield [
-            '{
-              "number": 1.4
-            }',
-            '{
+            'input' => '{ "number": 1.4 }',
+            'schema' => '{
               "type":"object",
               "properties":{
                 "number":{"type":"number"}
@@ -76,8 +70,8 @@ class NumberAndIntegerTypesTest extends BaseTestCase
             }'
         ];
         yield [
-            '{"number": 1e5}',
-            '{
+            'input' => '{"number": 1e5}',
+            'schema' => '{
                 "type": "object",
                 "properties": {
                     "number": {"type": "number"}
@@ -85,8 +79,8 @@ class NumberAndIntegerTypesTest extends BaseTestCase
             }'
         ];
         yield [
-            '{"number": 1}',
-            '{
+            'input' => '{"number": 1}',
+            'schema' => '{
                 "type": "object",
                 "properties": {
                     "number": {"type": "number"}
@@ -95,8 +89,8 @@ class NumberAndIntegerTypesTest extends BaseTestCase
             }'
         ];
         yield [
-            '{"number": -49.89}',
-            '{
+            'input' => '{"number": -49.89}',
+            'schema' => '{
                 "type": "object",
                 "properties": {
                     "number": {